backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.9
pydantic-settings==2.8.1
pypdf==4.2.0
python-docx==1.1.2
sqlalchemy==2.0.40
